+++ argostranslate/gui.py
@@ -101,7 +101,7 @@ class PackagesTable(QTableWidget):
         self.STRETCH_COLUMN_MIN_PADDING = 50
         self.horizontalHeader().setDefaultAlignment(Qt.AlignLeft)
 
-        self.available_packages = package.load_available_packages()
+        self.available_packages = package.get_available_packages()
         self.installed_packages = package.get_installed_packages()
 
     def populate(self):
@@ -161,7 +161,7 @@ class PackagesTable(QTableWidget):
 
     def get_packages(self):
         if self.table_content == self.TableContent.AVAILABLE:
-            return package.load_available_packages()
+            return package.get_available_packages()
         elif self.table_content == self.TableContent.INSTALLED:
             return package.get_installed_packages()
         else:
@@ -266,7 +266,7 @@ class DownloadPackagesWindow(QWidget):
         package.update_package_index()
 
         # Load available packages from local package index
-        available_packages = package.load_available_packages()
+        available_packages = package.get_available_packages()
 
         # Packages table
         self.packages_table = PackagesTable(PackagesTable.TableContent.AVAILABLE,